02-use-cases/healthcare-appointment-agent/requirements.txt
PyYAML==6.0.2
requests==2.32.4
requests-auth-aws-sigv4==0.7
python-dotenv==1.1.0
mcp[cli]>=1.10.0
strands-agents==0.1.8
ipykernel==6.29.5
aiohttp==3.12.14
langchain==0.3.27
langchain-aws==0.2.31
langchain-mcp-adapters==0.1.9
langgraph==0.6.6
botocore
bedrock-agentcore
boto3
langchain-google-genai==2.0.8
google-generativeai==0.8.3
